# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Local Office API
  version: 0.1.0
  summary: REST API for corporate lunch orchestration.
  description: |
    JSON-first API for managing lunch programs, orders, batching, delivery orchestration,
    invoicing, incidents, loyalty, and referrals. All mutating endpoints accept an
    Idempotency-Key header for safe retries and emit webhooks for state changes.
servers:
  - url: https://api.localoffice.example.com
    description: Production
  - url: https://staging-api.localoffice.example.com
    description: Staging
security:
  - bearerAuth: []
...
paths:
...
  /v1/orders:
    get:
      tags: [Orders]
      summary: List orders matching filters
      operationId: listOrders
      parameters:
        - name: org
          in: query
          schema:
            type: string
        - name: programSlotId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Order list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
    post:
      tags: [Orders]
      summary: Create an order
      description: |
        Creates a new order for an employee. Validates the 48-hour cutoff and enforces
        the maximum group size of 50 per batch. When subsidy applies, the total reflects
        discounts and loyalty credits.
      operationId: createOrder
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: Order violates cutoff or group size
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/orders/{id}/confirm:
    post:
      tags: [Orders]
      summary: Confirm an order and create payment intent
      operationId: confirmOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Order confirmed and payment initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderConfirmationResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Unique key to ensure idempotent processing of write operations.
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request body or parameters.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication credentials missing or invalid.
...
    NotFound:
      description: Resource not found.
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    CreateOrderRequest:
      type: object
      required: [programSlotId, userId, items]
      properties:
        programSlotId:
          type: string
        userId:
          type: string
        notes:
          type: string
        tip:
          type: number
          format: float
          minimum: 0
        gratuityPercentage:
          type: number
          format: float
          minimum: 0
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/OrderItemInput'
    OrderItemInput:
      type: object
      required: [skuId, quantity]
      properties:
        skuId:
          type: string
        quantity:
          type: integer
          minimum: 1
        modifiers:
          type: object
          additionalProperties: true
        notes:
          type: string
    OrderResponse:
      allOf:
        - $ref: '#/components/schemas/Order'
    OrderListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        nextCursor:
          type: string
          nullable: true
    OrderConfirmationResponse:
      type: object
      required: [order, payment]
      properties:
        order:
          $ref: '#/components/schemas/Order'
        payment:
          $ref: '#/components/schemas/PaymentIntent'
    Order:
      type: object
      required:
        - id
        - programSlotId
        - userId
        - status
        - subtotal
        - total
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        programSlotId:
          type: string
        userId:
          type: string
        batchId:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/OrderStatus'
        subtotal:
          type: number
          format: float
        tip:
          type: number
          format: float
        loyaltyDiscount:
          type: number
          format: float
        referralCredit:
          type: number
          format: float
        paymentFee:
          type: number
          format: float
        total:
          type: number
          format: float
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderItem:
      type: object
      required: [id, orderId, skuId, quantity]
      properties:
        id:
          type: string
        orderId:
          type: string
        skuId:
          type: string
        quantity:
          type: integer
        modifiers:
          type: object
          additionalProperties: true
        notes:
          type: string
    OrderStatus:
      type: string
      enum: [PENDING, LOCKED, BATCHED, FULFILLED, CANCELED]
    PaymentIntent:
      type: object
      required: [provider, clientSecret, methodOptions]
      properties:
        provider:
          type: string
          enum: [square]
        clientSecret:
          type: string
        methodOptions:
          type: object
          additionalProperties: true
...